requirements.txt
Flask==2.1.1
pdfminer.six==20221105
openai==0.27.0
python-dotenv==0.19.2